Minor but sparse refactoring, more in description
First of all, this refactoring was made to reach 100% code coverage, so I added some ignore comments for unreachable lines. Also, I changed possible return values for getCooldown function for CooldownStores to have an undefined return value: in fact, the undefined value should be used when no item is found, and the null value when it wasn't possible to get the cooldown because of other kind of reasons (like when the userId is not passed). I also added an error for when cooldownIdentifierKey is not passed for a RedisCooldownStore and an error for when options object in general is not passed to a RedisCooldownStore. Last, but not least, I simplified istanbul's job by rewriting some weird if statements.